+++ src/handlers/commands/editors/dry-run/dryRunEditor.ts
@@ -37,7 +37,7 @@ export function dryRunEditor<PARAMS extends EditOneOrAllParameters =
         repoLoader:
             p => gitHubRepoLoader(p.targets.credentials, DefaultDirectoryManager),
         editMode: ((params: PARAMS) => {
-            logger.info("About to create edit mode for dry run editor: params=%j", params);
+            logger.info("About to create edit mode for dry run editorCommand: params=%j", params);
             return new NewBranchWithStatus(
                 `edit-${name}-${Date.now()}`,
                 `${description.substr(0, 50)}\n\n[atomist] ${description}`,
