+++ src/software-delivery-machine/commands/editors/javaAffirmationEditor.ts
@@ -3,13 +3,13 @@ import { commitToMaster } from "@atomist/automation-client/operations/edit/editM
 import { SimpleProjectEditor } from "@atomist/automation-client/operations/edit/projectEditor";
 import { doWithFiles } from "@atomist/automation-client/project/util/projectUtils";
 import { AllJavaFiles } from "@atomist/spring-automation/commands/generator/java/javaProjectUtils";
-import { editor, EmptyParameters } from "../../../handlers/commands/editors/registerEditor";
+import { editorCommand, EmptyParameters } from "../../../handlers/commands/editors/editorCommand";
 
 /**
  * Harmlessly modify a Java file on master
  * @type {HandleCommand<EditOneOrAllParameters>}
  */
-export const javaAffirmationEditor: HandleCommand<any> = editor(
+export const javaAffirmationEditor: HandleCommand<any> = editorCommand(
     () => appendAffirmationToJava,
     "java affirmation",
     EmptyParameters,
@@ -22,7 +22,7 @@ export const javaAffirmationEditor: HandleCommand<any> = editor(
  * Harmlessly modify a Java file on a branch
  * @type {HandleCommand<EditOneOrAllParameters>}
  */
-export const javaBranchAffirmationEditor: HandleCommand<any> = editor(
+export const javaBranchAffirmationEditor: HandleCommand<any> = editorCommand(
     () => appendAffirmationToJava,
     "java branch affirmation",
     EmptyParameters,
