+++ src/software-delivery-machine/commands/editors/k8s/addK8sSpec.ts
@@ -1,7 +1,7 @@
 import { HandleCommand, logger } from "@atomist/automation-client";
 import { PullRequest } from "@atomist/automation-client/operations/edit/editModes";
 import { SimpleProjectEditor } from "@atomist/automation-client/operations/edit/projectEditor";
-import { editor, EmptyParameters } from "../../../../handlers/commands/editors/registerEditor";
+import { editorCommand, EmptyParameters } from "../../../../handlers/commands/editors/editorCommand";
 import { deploymentTemplate } from "./deploymentTemplate";
 
 export const AddK8sSpecCommandName = "AddK8sSpecCommandName";
@@ -10,7 +10,7 @@ export const AtomistK8sSpecFile = "atomist-k8s-deployment.json";
 
 export const ImagePlaceholder = "image_placeholder";
 
-export const addK8sSpec: HandleCommand<any> = editor(
+export const addK8sSpec: HandleCommand<any> = editorCommand(
     () => addK8sSpecEditor,
     AddK8sSpecCommandName,
     EmptyParameters,
