+++ src/software-delivery-machine/commands/generators/spring/springBootGenerator.ts
@@ -12,7 +12,7 @@ import { JavaGeneratorConfig } from "./JavaGeneratorConfig";
  * Function to create a Spring Boot generator.
  * Relies on generic Atomist Java & Spring functionality in spring-automations
  * @param config config for a Java generator, including location of seed
- * @param additionalActions zero or more additional editor actions
+ * @param additionalActions zero or more additional editorCommand actions
  * @return {HandleCommand<SpringBootGeneratorParameters>}
  */
 export function springBootGenerator(config: JavaGeneratorConfig,
@@ -55,13 +55,13 @@ export const setAtomistTeamInApplicationYml =
     };
 
 /**
- * Sample editor instance
+ * Sample editorCommand instance
  * @param {CustomSpringBootGeneratorParameters} params
  */
 export const sampleEditor: EditorFactory<CustomSpringBootGeneratorParameters> =
     params => async p => {
         /**
-         * Add any custom editor code to manipulate project ere
+         * Add any custom editorCommand code to manipulate project ere
          */
         return p;
     };
